Move some files into packages
Split up long files that were made of multiple parts
