+++ api/src/util/assert_never.py
