+++ api/tests/test_game_state_server.py
@@ -1,12 +1,12 @@
 import pytest
 
-from src.rate_limit import MemoryRateLimiterStorage, MemoryRateLimiter
+from src.rate_limit.memory_rate_limit import MemoryRateLimiterStorage, MemoryRateLimiter
 from tests.helpers import assert_matches, assert_all_match
 from src.game_state_server import GameStateServer, InvalidConnectionException
-from src.api_structures import CreateOrUpdateAction, DeleteAction, PingAction
-from src.room_store import MemoryRoomStore, MemoryRoomStorage
+from src.api.api_structures import CreateOrUpdateAction, DeleteAction, PingAction
+from src.room_store.memory_room_store import MemoryRoomStore, MemoryRoomStorage
 from src.game_components import Token, IconTokenContents
-from src.async_collect import async_collect
+from src.util.async_collect import async_collect
 from src.colors import colors
 from tests.fake_apm import fake_transaction
 from tests.static_fixtures import (
